Updated project to version 1.0.2

+++ chromepass/chrome_mac.py
@@ -19,11 +19,19 @@ class ChromeMac(Chrome):
 
         self.login_db_path = f"/Users/{getuser()}/Library/Application Support/Google/Chrome/Default/Login Data"
         self.tmp_login_db_path = f"/Users/{getuser()}/Library/Application Support/Google/Chrome/Default/Login_tmp"
-        shutil.copy2(self.login_db_path, self.tmp_login_db_path)  # making a temp copy since login data db is locked while chrome is running
+
+        if os.path.exists(self.login_db_path):
+            shutil.copy2(self.login_db_path, self.tmp_login_db_path)  # making a temp copy since login data db is locked while chrome is running
+        else:
+            print("It seems that no chrome browser is installed! Exiting...")
+            exit(1)
 
     def __del__(self):
         """destructor"""
-        os.remove(self.tmp_login_db_path)
+        try:
+            os.remove(self.tmp_login_db_path)
+        except FileNotFoundError as ex:
+            pass
 
     def decrypt_password(self, encrypted_password):
         """ decrypt the given encrypted password
